backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Google APIs
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Gemini AI
google-generativeai==0.8.3

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Session management
aiofiles==23.2.1
python-multipart==0.0.6
